Micro: Add Reservations functional component
